refactor: move EntityHalper to root refactor: graphic files to math
